refactor sia code into crate
