+++ mm2src/coins/Cargo.toml
@@ -18,7 +18,8 @@ enable-solana = [
 ]
 enable-sia = [
 	"dep:reqwest",
-	"blake2b_simd"
+	"blake2b_simd",
+	"dep:sia"
 ]
 default = []
 run-docker-tests = []
@@ -106,6 +107,7 @@ serde_json = { version = "1", features = ["preserve_order", "raw_value"] }
 serde_with = "1.14.0"
 serialization = { path = "../mm2_bitcoin/serialization" }
 serialization_derive = { path = "../mm2_bitcoin/serialization_derive" }
+sia = { path = "sia", optional = true }
 spv_validation = { path = "../mm2_bitcoin/spv_validation" }
 sha2 = "0.10"
 sha3 = "0.9"
