+++ mm2src/coins/sia/src/encoding.rs
@@ -1,5 +1,5 @@
-use crate::sia::blake2b_internal::hash_blake2b_single;
-use crate::sia::{PublicKey, Signature};
+use crate::blake2b_internal::hash_blake2b_single;
+use crate::{PublicKey, Signature};
 use rpc::v1::types::H256;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::convert::From;
