+++ mm2src/coins/sia/src/spend_policy.rs
@@ -1,8 +1,8 @@
-use crate::sia::address::Address;
-use crate::sia::blake2b_internal::{public_key_leaf, sigs_required_leaf, standard_unlock_hash, timelock_leaf,
+use crate::types::Address;
+use crate::blake2b_internal::{public_key_leaf, sigs_required_leaf, standard_unlock_hash, timelock_leaf,
                                    Accumulator};
-use crate::sia::encoding::{Encodable, Encoder, PrefixedH256, PrefixedPublicKey};
-use crate::sia::specifier::Specifier;
+use crate::encoding::{Encodable, Encoder, PrefixedH256, PrefixedPublicKey};
+use crate::specifier::Specifier;
 use ed25519_dalek::PublicKey;
 use nom::bytes::complete::{take_until, take_while, take_while_m_n};
 use nom::character::complete::char;
