+++ mm2src/coins/sia/src/transaction.rs
@@ -1,7 +1,6 @@
-use crate::sia::address::Address;
-use crate::sia::encoding::{Encodable, Encoder, HexArray64, PrefixedH256, PrefixedPublicKey, PrefixedSignature};
-use crate::sia::spend_policy::{SpendPolicy, SpendPolicyHelper, UnlockCondition, UnlockKey};
-use crate::sia::types::ChainIndex;
+use crate::encoding::{Encodable, Encoder, HexArray64, PrefixedH256, PrefixedPublicKey, PrefixedSignature};
+use crate::spend_policy::{SpendPolicy, SpendPolicyHelper, UnlockCondition, UnlockKey};
+use crate::types::{Address, ChainIndex};
 use ed25519_dalek::{PublicKey, Signature};
 use rpc::v1::types::H256;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
