+++ mm2src/coins_activation/src/prelude.rs
@@ -1,6 +1,6 @@
 use coins::nft::nft_structs::{Chain, ConvertChain};
 #[cfg(feature = "enable-sia")]
-use coins::sia::SiaCoinActivationParams;
+use coins::siacoin::SiaCoinActivationParams;
 use coins::utxo::UtxoActivationParams;
 use coins::z_coin::ZcoinActivationParams;
 use coins::{coin_conf, CoinBalance, CoinProtocol, DerivationMethodResponse, MmCoinEnum};
