+++ mm2src/mm2_main/src/rpc/dispatcher/dispatcher.rs
@@ -28,7 +28,7 @@ use coins::rpc_command::{account_balance::account_balance,
                          init_scan_for_new_addresses::{cancel_scan_for_new_addresses, init_scan_for_new_addresses,
                                                        init_scan_for_new_addresses_status},
                          init_withdraw::{cancel_withdraw, init_withdraw, withdraw_status, withdraw_user_action}};
-#[cfg(feature = "enable-sia")] use coins::sia::SiaCoin;
+#[cfg(feature = "enable-sia")] use coins::siacoin::SiaCoin;
 use coins::tendermint::{TendermintCoin, TendermintToken};
 use coins::utxo::bch::BchCoin;
 use coins::utxo::qtum::QtumCoin;
